data-model/src/main/python/requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements/requirements.txt setup.py
#
--trusted-host 192.168.2.110

certifi==2021.5.30
    # via requests
cfg-load==0.9.0
    # via cgnal-core (setup.py)
chardet==4.0.0
    # via requests
deprecated==1.2.12
    # via cgnal-core (setup.py)
dill==0.3.3
    # via cgnal-core (setup.py)
idna==2.10
    # via requests
mongomock==3.22.1
    # via cgnal-core (setup.py)
mpu[io]==0.23.1
    # via cfg-load
numpy==1.19.5
    # via
    #   cgnal-core (setup.py)
    #   pandas
pandas==1.1.5
    # via cgnal-core (setup.py)
pymongo==3.11.4
    # via cgnal-core (setup.py)
python-dateutil==2.8.1
    # via pandas
pytz==2021.1
    # via
    #   cfg-load
    #   mpu
    #   pandas
pyyaml==5.4.1
    # via
    #   cfg-load
    #   cgnal-core (setup.py)
requests==2.25.1
    # via
    #   cfg-load
    #   cgnal-core (setup.py)
sentinels==1.0.0
    # via mongomock
six==1.16.0
    # via
    #   cfg-load
    #   mongomock
    #   python-dateutil
tzlocal==2.1
    # via mpu
urllib3==1.26.5
    # via requests
wrapt==1.12.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
